+++ tests/test_docker_image_build.py
@@ -1,12 +1,12 @@
 import pytest
 
-from substratest.factory import DEFAULT_ALGO_METHOD_NAME
-from substratest.fl_interface import AlgoCategory
+from substratest.factory import DEFAULT_FUNCTION_NAME
+from substratest.fl_interface import FunctionCategory
 
 
 @pytest.mark.subprocess_skip
 def test_base_substra_tools_image(factory, cfg, client, default_dataset, worker):
-    """Test that an algo created with the base substra-tools image instead of the minimal works"""
+    """Test that an function created with the base substra-tools image instead of the minimal works"""
 
     suffix = "-minimal"
     if cfg.substra_tools.image_local.endswith(suffix):
@@ -14,19 +14,19 @@ def test_base_substra_tools_image(factory, cfg, client, default_dataset, worker)
     else:
         substra_tools_image = cfg.substra_tools.image_local
 
-    algo_category = AlgoCategory.simple
+    function_category = FunctionCategory.simple
 
     dockerfile = f"""
 FROM {substra_tools_image}
 
-COPY algo.py .
+COPY function.py .
 
-ENTRYPOINT ["python3", "algo.py", "--function-name", "{DEFAULT_ALGO_METHOD_NAME[algo_category]}"]
+ENTRYPOINT ["python3", "function.py", "--function-name", "{DEFAULT_FUNCTION_NAME[function_category]}"]
 """
-    spec = factory.create_algo(algo_category, dockerfile=dockerfile)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(function_category, dockerfile=dockerfile)
+    function = client.add_function(spec)
     spec = factory.create_traintask(
-        algo=algo,
+        function=function,
         inputs=default_dataset.train_data_inputs,
         worker=worker,
     )