+++ tests/test_execution.py
@@ -4,8 +4,8 @@ from substra.sdk.models import Status
 from substra.sdk.schemas import TaskSpec
 
 import substratest as sbt
-from substratest.factory import AlgoCategory
 from substratest.factory import AugmentedDataset
+from substratest.factory import FunctionCategory
 from substratest.fl_interface import FLTaskInputGenerator
 from substratest.fl_interface import FLTaskOutputGenerator
 from substratest.fl_interface import InputIdentifiers
@@ -16,16 +16,16 @@ from substratest.fl_interface import OutputIdentifiers
 def test_tasks_execution_on_same_organization(factory, network, client, default_dataset, default_metric, worker):
     """Execution of a traintask, a following testtask and a following traintask."""
 
-    spec = factory.create_algo(AlgoCategory.simple)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.simple)
+    function = client.add_function(spec)
 
-    predict_algo_spec = factory.create_algo(AlgoCategory.predict)
-    predict_algo = client.add_algo(predict_algo_spec)
+    predict_function_spec = factory.create_function(FunctionCategory.predict)
+    predict_function = client.add_function(predict_function_spec)
 
     # create traintask
     def get_traintask_spec() -> TaskSpec:
         return factory.create_traintask(
-            algo=algo,
+            function=function,
             inputs=default_dataset.train_data_inputs,
             metadata={"foo": "bar"},
             worker=worker,
@@ -50,7 +50,7 @@ def test_tasks_execution_on_same_organization(factory, network, client, default_
 
     # create testtask
     spec = factory.create_predicttask(
-        algo=predict_algo,
+        function=predict_function,
         inputs=default_dataset.test_data_inputs + FLTaskInputGenerator.train_to_predict(traintask.key),
         worker=worker,
     )
@@ -61,7 +61,7 @@ def test_tasks_execution_on_same_organization(factory, network, client, default_
     assert predicttask.error_type is None
 
     spec = factory.create_testtask(
-        algo=default_metric,
+        function=default_metric,
         inputs=default_dataset.test_data_inputs + FLTaskInputGenerator.predict_to_test(predicttask.key),
         worker=worker,
     )
@@ -74,7 +74,7 @@ def test_tasks_execution_on_same_organization(factory, network, client, default_
     # add a traintask depending on first traintask
     first_traintask_key = traintask.key
     spec = factory.create_traintask(
-        algo=algo,
+        function=function,
         inputs=default_dataset.train_data_inputs + FLTaskInputGenerator.trains_to_train([first_traintask_key]),
         metadata=None,
         worker=worker,
@@ -94,8 +94,8 @@ def test_federated_learning_workflow(factory, client, default_datasets, workers)
     """Test federated learning workflow on each organization."""
 
     # create test environment
-    spec = factory.create_algo(AlgoCategory.simple)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.simple)
+    function = client.add_function(spec)
 
     # create 1 traintask per dataset and chain them
     traintask = None
@@ -105,11 +105,10 @@ def test_federated_learning_workflow(factory, client, default_datasets, workers)
     # default_datasets contains datasets on each organization and
     # that has a result we can use for federated learning
     for index, dataset in enumerate(default_datasets):
-
         traintasks = [traintask.key] if traintask else []
 
         spec = factory.create_traintask(
-            algo=algo,
+            function=function,
             inputs=dataset.train_data_inputs + FLTaskInputGenerator.trains_to_train(traintasks),
             tag="foo",
             rank=rank,
@@ -147,18 +146,18 @@ def test_tasks_execution_on_different_organizations(
 ):
     """Execution of a traintask on organization 1 and the following testtask on organization 2."""
     # add test data samples / dataset / metric on organization 1
-    spec = factory.create_algo(AlgoCategory.simple)
-    algo_2 = client_2.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.simple)
+    function_2 = client_2.add_function(spec)
 
-    predict_algo_spec = factory.create_algo(AlgoCategory.predict)
-    predict_algo_2 = client_2.add_algo(predict_algo_spec)
+    predict_function_spec = factory.create_function(FunctionCategory.predict)
+    predict_function_2 = client_2.add_function(predict_function_spec)
 
-    channel.wait_for_asset_synchronized(algo_2)
-    channel.wait_for_asset_synchronized(predict_algo_2)
+    channel.wait_for_asset_synchronized(function_2)
+    channel.wait_for_asset_synchronized(predict_function_2)
 
     # add traintask on organization 2; should execute on organization 2 (dataset located on organization 2)
     spec = factory.create_traintask(
-        algo=algo_2,
+        function=function_2,
         inputs=default_dataset_2.train_data_inputs,
         worker=workers[1],
     )
@@ -172,7 +171,7 @@ def test_tasks_execution_on_different_organizations(
 
     # add testtask; should execute on organization 1 (default_dataset_1 is located on organization 1)
     spec = factory.create_predicttask(
-        algo=predict_algo_2,
+        function=predict_function_2,
         inputs=default_dataset_1.test_data_inputs + FLTaskInputGenerator.train_to_predict(traintask.key),
         worker=workers[0],
     )
@@ -183,7 +182,7 @@ def test_tasks_execution_on_different_organizations(
     assert predicttask.worker == client_1.organization_id
 
     spec = factory.create_testtask(
-        algo=default_metric_1,
+        function=default_metric_1,
         inputs=default_dataset_1.test_data_inputs + FLTaskInputGenerator.predict_to_test(predicttask.key),
         worker=workers[0],
     )
@@ -197,15 +196,17 @@ def test_tasks_execution_on_different_organizations(
 
 @pytest.mark.slow
 @pytest.mark.subprocess_skip
-def test_algo_build_failure(factory, network, default_dataset_1, worker):
+def test_function_build_failure(factory, network, default_dataset_1, worker):
     """Invalid Dockerfile is causing compute task failure."""
 
-    dockerfile = factory.default_algo_dockerfile(method_name=sbt.factory.DEFAULT_ALGO_METHOD_NAME[AlgoCategory.simple])
+    dockerfile = factory.default_function_dockerfile(
+        method_name=sbt.factory.DEFAULT_FUNCTION_NAME[FunctionCategory.simple]
+    )
     dockerfile += "\nRUN invalid_command"
-    spec = factory.create_algo(category=AlgoCategory.simple, dockerfile=dockerfile)
-    algo = network.clients[0].add_algo(spec)
+    spec = factory.create_function(category=FunctionCategory.simple, dockerfile=dockerfile)
+    function = network.clients[0].add_function(spec)
 
-    spec = factory.create_traintask(algo=algo, inputs=default_dataset_1.train_data_inputs, worker=worker)
+    spec = factory.create_traintask(function=function, inputs=default_dataset_1.train_data_inputs, worker=worker)
 
     if network.clients[0].backend_mode != substra.BackendType.REMOTE:
         with pytest.raises(substra.sdk.backends.local.compute.spawner.base.BuildError):
@@ -226,12 +227,12 @@ def test_algo_build_failure(factory, network, default_dataset_1, worker):
 
 @pytest.mark.slow
 def test_task_execution_failure(factory, network, default_dataset_1, worker):
-    """Invalid algo script is causing compute task failure."""
+    """Invalid function script is causing compute task failure."""
 
-    spec = factory.create_algo(category=AlgoCategory.simple, py_script=sbt.factory.INVALID_ALGO_SCRIPT)
-    algo = network.clients[0].add_algo(spec)
+    spec = factory.create_function(category=FunctionCategory.simple, py_script=sbt.factory.INVALID_FUNCTION_SCRIPT)
+    function = network.clients[0].add_function(spec)
 
-    spec = factory.create_traintask(algo=algo, inputs=default_dataset_1.train_data_inputs, worker=worker)
+    spec = factory.create_traintask(function=function, inputs=default_dataset_1.train_data_inputs, worker=worker)
 
     if network.clients[0].backend_mode != substra.BackendType.REMOTE:
         with pytest.raises(substra.sdk.backends.local.compute.spawner.base.ExecutionError):
@@ -252,12 +253,14 @@ def test_task_execution_failure(factory, network, default_dataset_1, worker):
 
 @pytest.mark.slow
 def test_composite_traintask_execution_failure(factory, client, default_dataset, worker):
-    """Invalid composite algo script is causing traintask failure."""
+    """Invalid composite function script is causing traintask failure."""
 
-    spec = factory.create_algo(AlgoCategory.composite, py_script=sbt.factory.INVALID_COMPOSITE_ALGO_SCRIPT)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.composite, py_script=sbt.factory.INVALID_COMPOSITE_FUNCTION_SCRIPT)
+    function = client.add_function(spec)
 
-    spec = factory.create_composite_traintask(algo=algo, inputs=default_dataset.train_data_inputs, worker=worker)
+    spec = factory.create_composite_traintask(
+        function=function, inputs=default_dataset.train_data_inputs, worker=worker
+    )
     if client.backend_mode == substra.BackendType.REMOTE:
         composite_traintask = client.add_task(spec)
         composite_traintask = client.wait(composite_traintask, raises=False)
@@ -278,25 +281,25 @@ def test_composite_traintask_execution_failure(factory, client, default_dataset,
 
 @pytest.mark.slow
 def test_aggregatetask_execution_failure(factory, client, default_dataset, worker):
-    """Invalid algo script is causing traintask failure."""
+    """Invalid function script is causing traintask failure."""
 
-    spec = factory.create_algo(AlgoCategory.composite)
-    composite_algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.composite)
+    composite_function = client.add_function(spec)
 
-    spec = factory.create_algo(AlgoCategory.aggregate, py_script=sbt.factory.INVALID_AGGREGATE_ALGO_SCRIPT)
-    aggregate_algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.aggregate, py_script=sbt.factory.INVALID_AGGREGATE_FUNCTION_SCRIPT)
+    aggregate_function = client.add_function(spec)
 
     composite_traintask_keys = []
     for i in [0, 1]:
         spec = factory.create_composite_traintask(
-            algo=composite_algo,
+            function=composite_function,
             inputs=default_dataset.opener_input + [default_dataset.train_data_sample_inputs[i]],
             worker=worker,
         )
         composite_traintask_keys.append(client.add_task(spec).key)
 
     spec = factory.create_aggregatetask(
-        algo=aggregate_algo,
+        function=aggregate_function,
         inputs=FLTaskInputGenerator.composites_to_aggregate(composite_traintask_keys),
         worker=client.organization_id,
     )
@@ -327,15 +330,15 @@ def test_aggregatetask_execution_failure(factory, client, default_dataset, worke
 def test_composite_traintasks_execution(factory, client, default_dataset, default_metric, worker):
     """Execution of composite traintasks."""
 
-    spec = factory.create_algo(AlgoCategory.composite)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.composite)
+    function = client.add_function(spec)
 
-    spec = factory.create_algo(AlgoCategory.predict_composite)
-    predict_algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.predict_composite)
+    predict_function = client.add_function(spec)
 
     # first composite traintask
     spec = factory.create_composite_traintask(
-        algo=algo,
+        function=function,
         inputs=default_dataset.train_data_inputs,
         worker=worker,
     )
@@ -347,7 +350,7 @@ def test_composite_traintasks_execution(factory, client, default_dataset, defaul
 
     # second composite traintask
     spec = factory.create_composite_traintask(
-        algo=algo,
+        function=function,
         inputs=default_dataset.train_data_inputs
         + FLTaskInputGenerator.composite_to_composite(composite_traintask_1.key),
         worker=worker,
@@ -360,7 +363,7 @@ def test_composite_traintasks_execution(factory, client, default_dataset, defaul
 
     # add a 'composite' testtask
     spec = factory.create_predicttask(
-        algo=predict_algo,
+        function=predict_function,
         inputs=default_dataset.test_data_inputs + FLTaskInputGenerator.composite_to_predict(composite_traintask_2.key),
         worker=worker,
     )
@@ -370,7 +373,7 @@ def test_composite_traintasks_execution(factory, client, default_dataset, defaul
     assert predicttask.error_type is None
 
     spec = factory.create_testtask(
-        algo=default_metric,
+        function=default_metric,
         inputs=default_dataset.test_data_inputs + FLTaskInputGenerator.predict_to_test(predicttask.key),
         worker=worker,
     )
@@ -394,26 +397,26 @@ def test_aggregatetask(factory, client, default_metric, default_dataset, worker)
 
     train_data_sample_inputs = default_dataset.train_data_sample_inputs[:number_of_traintasks_to_aggregate]
 
-    spec = factory.create_algo(AlgoCategory.simple)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.simple)
+    function = client.add_function(spec)
 
-    spec = factory.create_algo(AlgoCategory.predict)
-    predict_algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.predict)
+    predict_function = client.add_function(spec)
 
     # add traintasks
     traintask_keys = []
     for data_sample_input in train_data_sample_inputs:
         spec = factory.create_traintask(
-            algo=algo, inputs=default_dataset.opener_input + [data_sample_input], worker=worker
+            function=function, inputs=default_dataset.opener_input + [data_sample_input], worker=worker
         )
         traintask = client.add_task(spec)
         traintask_keys.append(traintask.key)
 
-    spec = factory.create_algo(AlgoCategory.aggregate)
-    aggregate_algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.aggregate)
+    aggregate_function = client.add_function(spec)
 
     spec = factory.create_aggregatetask(
-        algo=aggregate_algo,
+        function=aggregate_function,
         worker=client.organization_id,
         inputs=FLTaskInputGenerator.trains_to_aggregate(traintask_keys),
     )
@@ -424,7 +427,7 @@ def test_aggregatetask(factory, client, default_metric, default_dataset, worker)
     )
 
     spec = factory.create_predicttask(
-        algo=predict_algo,
+        function=predict_function,
         inputs=default_dataset.test_data_inputs + FLTaskInputGenerator.aggregate_to_predict(aggregatetask.key),
         worker=worker,
     )
@@ -432,7 +435,7 @@ def test_aggregatetask(factory, client, default_metric, default_dataset, worker)
     predicttask = client.wait(predicttask)
 
     spec = factory.create_testtask(
-        algo=default_metric,
+        function=default_metric,
         inputs=default_dataset.test_data_inputs + FLTaskInputGenerator.predict_to_test(predicttask.key),
         worker=worker,
     )
@@ -448,23 +451,23 @@ def test_aggregatetask_chained(factory, client, default_dataset, worker):
 
     train_data_sample_input = default_dataset.train_data_sample_inputs[:1]
 
-    spec = factory.create_algo(AlgoCategory.simple)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.simple)
+    function = client.add_function(spec)
 
     # add traintasks
     spec = factory.create_traintask(
-        algo=algo,
+        function=function,
         inputs=default_dataset.opener_input + train_data_sample_input,
         worker=worker,
     )
     traintask = client.add_task(spec)
 
-    spec = factory.create_algo(AlgoCategory.aggregate)
-    aggregate_algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.aggregate)
+    aggregate_function = client.add_function(spec)
 
     # add first layer of aggregatetasks
     spec = factory.create_aggregatetask(
-        algo=aggregate_algo,
+        function=aggregate_function,
         worker=client.organization_id,
         inputs=FLTaskInputGenerator.trains_to_aggregate([traintask.key]),
     )
@@ -477,7 +480,7 @@ def test_aggregatetask_chained(factory, client, default_dataset, worker):
 
     # add second layer of aggregatetask
     spec = factory.create_aggregatetask(
-        algo=aggregate_algo,
+        function=aggregate_function,
         worker=client.organization_id,
         inputs=FLTaskInputGenerator.trains_to_aggregate([aggregatetask_1.key]),
     )
@@ -499,23 +502,23 @@ def test_aggregatetask_traintask(factory, client, default_dataset, worker):
     train_data_sample_input_1 = train_data_sample_inputs[0]
     train_data_sample_input_2 = train_data_sample_inputs[1]
 
-    spec = factory.create_algo(AlgoCategory.simple)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.simple)
+    function = client.add_function(spec)
 
     # add first part of the traintasks
     spec = factory.create_traintask(
-        algo=algo,
+        function=function,
         inputs=default_dataset.opener_input + [train_data_sample_input_1],
         worker=worker,
     )
     traintask_1 = client.add_task(spec)
 
-    spec = factory.create_algo(AlgoCategory.aggregate)
-    aggregate_algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.aggregate)
+    aggregate_function = client.add_function(spec)
 
     # add aggregatetask
     spec = factory.create_aggregatetask(
-        algo=aggregate_algo,
+        function=aggregate_function,
         worker=client.organization_id,
         inputs=FLTaskInputGenerator.trains_to_aggregate([traintask_1.key]),
     )
@@ -524,7 +527,7 @@ def test_aggregatetask_traintask(factory, client, default_dataset, worker):
 
     # add second part of the traintasks
     spec = factory.create_traintask(
-        algo=algo,
+        function=function,
         inputs=default_dataset.opener_input
         + [train_data_sample_input_2]
         + FLTaskInputGenerator.trains_to_train([aggregatetask.key]),
@@ -545,14 +548,14 @@ def test_composite_traintask_2_organizations_to_composite_traintask(factory, cli
     organization 1 and another composite traintask (inpute_trunk_model) from organization 2
     """
 
-    spec = factory.create_algo(AlgoCategory.composite)
-    composite_algo = clients[0].add_algo(spec)
+    spec = factory.create_function(FunctionCategory.composite)
+    composite_function = clients[0].add_function(spec)
 
     # composite traintasks on organization 1 and organization 2
     composite_traintask_keys = []
     for index, dataset in enumerate(default_datasets):
         spec = factory.create_composite_traintask(
-            algo=composite_algo,
+            function=composite_function,
             inputs=dataset.opener_input + dataset.train_data_sample_inputs[:1],
             outputs=FLTaskOutputGenerator.composite_traintask(
                 shared_authorized_ids=[c.organization_id for c in clients],
@@ -564,7 +567,7 @@ def test_composite_traintask_2_organizations_to_composite_traintask(factory, cli
         composite_traintask_keys.append(composite_traintask_key)
 
     spec = factory.create_composite_traintask(
-        algo=composite_algo,
+        function=composite_function,
         inputs=default_datasets[0].train_data_inputs
         + FLTaskInputGenerator.composite_to_composite(composite_traintask_keys[0], composite_traintask_keys[1]),
         rank=1,
@@ -613,15 +616,15 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
     aggregate_worker = clients[0].organization_id
     number_of_rounds = 2
 
-    # register algos on first organization
-    spec = factory.create_algo(AlgoCategory.composite)
-    composite_algo = clients[0].add_algo(spec)
-    spec = factory.create_algo(AlgoCategory.aggregate)
-    aggregate_algo = clients[0].add_algo(spec)
-    spec = factory.create_algo(AlgoCategory.predict)
-    predict_algo = clients[0].add_algo(spec)
-    spec = factory.create_algo(AlgoCategory.predict_composite)
-    predict_algo_composite = clients[0].add_algo(spec)
+    # register functions on first organization
+    spec = factory.create_function(FunctionCategory.composite)
+    composite_function = clients[0].add_function(spec)
+    spec = factory.create_function(FunctionCategory.aggregate)
+    aggregate_function = clients[0].add_function(spec)
+    spec = factory.create_function(FunctionCategory.predict)
+    predict_function = clients[0].add_function(spec)
+    spec = factory.create_function(FunctionCategory.predict_composite)
+    predict_function_composite = clients[0].add_function(spec)
 
     # launch execution
     previous_aggregatetask_key = None
@@ -631,7 +634,6 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
         # create composite traintask on each organization
         composite_traintask_keys = []
         for index, dataset in enumerate(default_datasets):
-
             if previous_aggregatetask_key:
                 input_models = FLTaskInputGenerator.composite_to_local(
                     previous_composite_traintask_keys[index]
@@ -641,7 +643,7 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
                 input_models = []
 
             spec = factory.create_composite_traintask(
-                algo=composite_algo,
+                function=composite_function,
                 inputs=[dataset.train_data_sample_inputs[0 + round_]] + dataset.opener_input + input_models,
                 outputs=FLTaskOutputGenerator.composite_traintask(
                     shared_authorized_ids=[c.organization_id for c in clients],
@@ -656,7 +658,7 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
 
         # create aggregate on its organization
         spec = factory.create_aggregatetask(
-            algo=aggregate_algo,
+            function=aggregate_function,
             worker=aggregate_worker,
             inputs=FLTaskInputGenerator.composites_to_aggregate(composite_traintask_keys),
         )
@@ -672,7 +674,7 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
         zip(previous_composite_traintask_keys, default_metrics, default_datasets)
     ):
         spec = factory.create_predicttask(
-            algo=predict_algo_composite,
+            function=predict_function_composite,
             inputs=dataset.test_data_inputs + FLTaskInputGenerator.composite_to_predict(traintask_key),
             worker=workers[index],
         )
@@ -680,7 +682,7 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
         predicttask = clients[0].wait(predicttask)
 
         spec = factory.create_testtask(
-            algo=metric,
+            function=metric,
             inputs=dataset.test_data_inputs + FLTaskInputGenerator.predict_to_test(predicttask.key),
             worker=workers[index],
         )
@@ -690,7 +692,7 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
         assert testtask.outputs[OutputIdentifiers.performance].value == pytest.approx(32 + index)
 
     spec = factory.create_predicttask(
-        algo=predict_algo,
+        function=predict_function,
         inputs=default_datasets[0].test_data_inputs
         + FLTaskInputGenerator.aggregate_to_predict(previous_aggregatetask_key),
         worker=workers[0],
@@ -699,7 +701,7 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
     predicttask = clients[0].wait(predicttask)
 
     spec = factory.create_testtask(
-        algo=default_metrics[0],
+        function=default_metrics[0],
         inputs=default_datasets[0].test_data_inputs + FLTaskInputGenerator.predict_to_test(predicttask.key),
         worker=workers[0],
     )
@@ -730,9 +732,9 @@ def test_aggregate_composite_traintasks(factory, network, clients, default_datas
 
         client = clients[0]
         dataset = default_datasets[0]
-        algo = client.add_algo(spec)
+        function = client.add_function(spec)
 
-        spec = factory.create_traintask(algo=algo, inputs=dataset.train_data_inputs, worker=workers[0])
+        spec = factory.create_traintask(function=function, inputs=dataset.train_data_inputs, worker=workers[0])
         traintask = client.add_task(spec)
         traintask = client.wait(traintask)
         assert traintask.status == Status.failed
@@ -756,13 +758,13 @@ def test_use_data_sample_located_in_shared_path(factory, network, client, organi
         train_data_sample_keys=[data_sample_key],
     )
 
-    spec = factory.create_algo(AlgoCategory.simple)
-    algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.simple)
+    function = client.add_function(spec)
 
-    spec = factory.create_algo(AlgoCategory.predict)
-    predict_algo = client.add_algo(spec)
+    spec = factory.create_function(FunctionCategory.predict)
+    predict_function = client.add_function(spec)
 
-    spec = factory.create_traintask(algo=algo, inputs=dataset.train_data_inputs, worker=worker)
+    spec = factory.create_traintask(function=function, inputs=dataset.train_data_inputs, worker=worker)
     traintask = client.add_task(spec)
     traintask = client.wait(traintask)
     assert traintask.status == Status.done
@@ -771,7 +773,7 @@ def test_use_data_sample_located_in_shared_path(factory, network, client, organi
 
     # create testtask
     spec = factory.create_predicttask(
-        algo=predict_algo, traintask=traintask, dataset=dataset, data_samples=[data_sample_key], worker=worker
+        function=predict_function, traintask=traintask, dataset=dataset, data_samples=[data_sample_key], worker=worker
     )
     predicttask = client.add_task(spec)
     predicttask = client.wait(predicttask)
@@ -779,7 +781,7 @@ def test_use_data_sample_located_in_shared_path(factory, network, client, organi
     assert predicttask.error_type is None
 
     spec = factory.create_testtask(
-        algo=default_metric, predicttask=predicttask, dataset=dataset, data_samples=[data_sample_key], worker=worker
+        function=default_metric, predicttask=predicttask, dataset=dataset, data_samples=[data_sample_key], worker=worker
     )
     testtask = client.add_task(spec)
     testtask = client.wait(testtask)
@@ -792,10 +794,11 @@ def test_use_data_sample_located_in_shared_path(factory, network, client, organi
 def test_user_creates_model_folder(factory, client, default_dataset, worker):
     """Check that the model folder is not overwritten by substra"""
     dockerfile = (
-        f"FROM {factory.default_tools_image}\nCOPY algo.py .\nRUN mkdir model\n"
-        + 'RUN echo \'{"name":"Jane"}\' >> model/model\nENTRYPOINT ["python3", "algo.py", "--function-name", "train"]\n'
+        f"FROM {factory.default_tools_image}\nCOPY function.py .\nRUN mkdir model\n"
+        + 'RUN echo \'{"name":"Jane"}\' >> model/model'
+        + '\nENTRYPOINT ["python3", "function.py", "--function-name", "train"]\n'
     )
-    algo_script = f"""
+    function_script = f"""
 import json
 import substratools as tools
 
@@ -829,14 +832,14 @@ def save_predictions(predictions, path):
 if __name__ == '__main__':
     tools.execute()
 """  # noqa
-    spec = factory.create_algo(AlgoCategory.simple, py_script=algo_script, dockerfile=dockerfile)
-    algo = client.add_algo(spec)
-    spec = factory.create_traintask(algo=algo, inputs=default_dataset.train_data_inputs, worker=worker)
+    spec = factory.create_function(FunctionCategory.simple, py_script=function_script, dockerfile=dockerfile)
+    function = client.add_function(spec)
+    spec = factory.create_traintask(function=function, inputs=default_dataset.train_data_inputs, worker=worker)
     traintask = client.add_task(spec)
     client.wait(traintask)
 
 
-WRITE_TO_HOME_DIRECTORY_ALGO = f"""
+WRITE_TO_HOME_DIRECTORY_FUNCTION = f"""
 import json
 import substratools as tools
 
@@ -878,11 +881,11 @@ if __name__ == '__main__':
 
 @pytest.mark.subprocess_skip
 def test_write_to_home_directory(factory, client, default_dataset, worker):
-    """The algo writes to the home directory (~/foo)"""
+    """The function writes to the home directory (~/foo)"""
 
-    spec = factory.create_algo(AlgoCategory.simple, WRITE_TO_HOME_DIRECTORY_ALGO)
-    algo = client.add_algo(spec)
-    spec = factory.create_traintask(algo=algo, inputs=default_dataset.train_data_inputs, worker=worker)
+    spec = factory.create_function(FunctionCategory.simple, WRITE_TO_HOME_DIRECTORY_FUNCTION)
+    function = client.add_function(spec)
+    spec = factory.create_traintask(function=function, inputs=default_dataset.train_data_inputs, worker=worker)
     traintask = client.add_task(spec)
     traintask = client.wait(traintask)