+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- BREAKING: rename Algo to Function (#243)
+
 ## [0.38.1] - 2023-02-06
 
 ### Added